add a plot output as vector graphics into a Word object

produces a vector graphics output from R plot instructions and add the result in an Word document object produced by read_docx.

body_add_vg(x, code, pos = "after", ...)

Arguments

x
an rdocx object produced by officer::read_docx

code
plot instructions

pos
where to add the new element relative to the cursor, one of "after", "before", "on".

...
arguments passed on to dml_docx.

Note

The function is maintained but using it should be avoided: Word text boxes, the elements used to put text in a graphic, are adding extra space on top and bottom of the shape. As there is no clear rule available to handle that, it makes impossible to compute what should be the exact position of a text. This can affect the whole rendering of the graphic.

The function should then be considered as a failed experience. An alternative is to use EMF format, this will not allow editing the graphic but the display is made as vector graphic.

Examples

# \donttest{
library(officer)
x <- read_docx()
x <- body_add_vg(x, code = barplot(1:5, col = 2:6) )

#> Warning: dml_docx() is deprecated and will be removed in the next version (> 0.2.0), use devEMF::emf instead.

print(x, target = "vg.docx")
# }