DrawingML graphic device for Microsoft PowerPoint

Graphics devices for Microsoft PowerPoint DrawingML format.

dml_pptx(file = "Rplots.dml", width = 6, height = 6, offx = 1,
  offy = 1, bg = "white", fonts = list(), pointsize = 12,
  editable = TRUE, id = 1L, last_rel_id = 1L,
  raster_prefix = "raster_", standalone = TRUE)

Arguments

file
the file where output will appear.

height, width
Height and width in inches.

offx, offy
top and left origin of the plot

bg
Default background color for the plot (defaults to "white").

fonts
Named list of font names to be aliased with fonts installed on your system. If unspecified, the R default families sans, serif, mono and symbol are aliased to the family returned by match_family().

When you use specific fonts, you will need that font installed on your system. This can be check with package gdtools and function gdtools::font_family_exists().

An example: list(sans = "Roboto", serif = "Times", mono = "Courier").

pointsize
default point size.

editable
should vector graphics elements (points, text, etc.) be editable.

id
specifies a unique identifier (integer) within the slide that will contain the DrawingML instructions.

last_rel_id
specifies the last unique identifier (integer) within relationship file that will be used to reference embedded raster images if any.

raster_prefix
string value used as prefix for png files produced when raster objects are printed on the graphical device.

standalone
produce a standalone drawingml file? If FALSE, omits xml header and namespaces.

See also

Examples

# \donttest{
dml_pptx( file = tempfile() )
plot(1:11,(-5:5)^2, type='b', main="Simple Example")
dev.off()

#> pdf 
#>   2 

# }